pyproject.toml
[tool.poetry]
name = "enrichr-kg-builder"
version = "0.1.0"
description = "Knowledge Graph Builder for Enrichr data using BioCypher"
authors = ["Your Name <your.email@example.com>"]
readme = "README_ENRICHR.md"

[tool.poetry.dependencies]
python = "^3.10"
biocypher = "^0.5.4"
requests = "^2.31.0"
treelib = "1.6.4"
numpy = "^1.26"
rdflib = "^6.2.0"  
openpyxl = "^3.1.5"
pandas = "^2.0.0"
pyyaml = "^6.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
enrichr-kg-builder = "enrichr_kg_builder:main"
